perf: Update minimap2 settings for nanopore data
This commit should address the problems with deletions that happen in
overlap regions of reads, or rather at the end of reads without primers
(e.g. a common deltion that happens in Lambda variant).

The main change is addition of an end-bonus, which is justified as the
primers are removed in the rule before the alignment. The other big
change is a reduction of the gap extension penalty and an increase in
the match and mismatch costs.